+++ hybrasyl/Objects/User.cs
@@ -36,23 +36,6 @@ namespace Hybrasyl.Objects
 {
 
     [JsonObject]
-    public class LegendMark
-    {
-        public String Prefix { get; set; }
-        public int Color { get; set; }
-        public int Icon { get; set; }
-        public String Text { get; set; }
-        public bool Public { get; set; }
-        public DateTime Created { get; set; }
-        public int Quantity { get; set; }
-        /*
-        public override string ToString()
-        {
-            //var ingame = World.TimeConverter.AsString(Created, bool fullTime = false)
-        }*/
-    }
-
-    [JsonObject]
     public class GuildMembership
     {
         public String Title { get; set; }
@@ -168,10 +151,18 @@ namespace Hybrasyl.Objects
             {
                 return Nation != null ? Nation.Name : string.Empty;
             }
-        } 
-        [JsonProperty]
-        public List<LegendMark> Legend { get; set; }
+        }
 
+        [JsonProperty] public Legend Legend;
+
+        /// <summary>
+        /// Reindexes any temporary data structures that may need to be recreated after a user is deserialized from JSON data.
+        /// </summary>
+        public void Reindex()
+        {
+            Legend.RegenerateIndex();    
+        }
+            
         public DialogState DialogState { get; set; }
 
         [JsonProperty]
@@ -240,6 +231,10 @@ namespace Hybrasyl.Objects
                 return span.TotalSeconds < 0 ? 0 : span.TotalSeconds;
             }
         }
+
+        public string SinceLastLoginString => SinceLastLogin < 86400 ? 
+            $"{Math.Floor(SinceLastLogin/3600)} hours, {Math.Floor(SinceLastLogin%3600/60)} minutes" : 
+            $"{Math.Floor(SinceLastLogin/86400)} days, {Math.Floor(SinceLastLogin%86400/3600)} hours, {Math.Floor(SinceLastLogin%86400%3600)/60} minutes";
 
         // Throttling checks for messaging
 
@@ -352,7 +347,7 @@ namespace Hybrasyl.Objects
             Login = new LoginInfo();
             Password = new PasswordInfo();
             Location = new Location();
-            Legend = new List<LegendMark>();
+            Legend = new Legend();
             Guild = new GuildMembership();
             LastSaid = String.Empty;
             LastSpoke = 0;
@@ -699,8 +694,8 @@ namespace Hybrasyl.Objects
             profilePacket.WriteString8(Guild.Rank);
             profilePacket.WriteString8(Hybrasyl.Constants.REVERSE_CLASSES[(int)Class]);
             profilePacket.WriteString8(Guild.Name);
-            profilePacket.WriteByte((byte)Legend.Count );
-            foreach (var mark in Legend)
+            profilePacket.WriteByte((byte)Legend.Count);
+            foreach (var mark in Legend.Where(mark => mark.Public))
             {
                 profilePacket.WriteByte((byte)mark.Icon);
                 profilePacket.WriteByte((byte)mark.Color);
@@ -1800,10 +1795,10 @@ namespace Hybrasyl.Objects
             profilePacket.WriteByte((byte) Legend.Count);
             foreach (var mark in Legend)
             {
-                profilePacket.WriteByte((byte) mark.Icon);
-                profilePacket.WriteByte((byte) mark.Color);
+                profilePacket.WriteByte((byte)mark.Icon);
+                profilePacket.WriteByte((byte)mark.Color);
                 profilePacket.WriteString8(mark.Prefix);
-                profilePacket.WriteString8(mark.Text);
+                profilePacket.WriteString8(mark.ToString());
             }
 
             Enqueue(profilePacket);
@@ -2308,51 +2303,5 @@ namespace Hybrasyl.Objects
             Client.SendMessage(p, 3);
         }
 
-
-        /*
-        [SecurityPermission(SecurityAction.Demand, SerializationFormatter =true)]
-        public void GetObjectData(SerializationInfo info, StreamingContext context)
-        {
-            info.AddValue("SerializationVersion", "1");
-            info.AddValue("Name", Name);
-            info.AddValue("Sex", Sex);
-            info.AddValue("HairStyle", HairStyle);
-            info.AddValue("HairColor", HairColor);
-            info.AddValue("Class", Class);
-            info.AddValue("Level", Level);
-            info.AddValue("LevelPoints", LevelPoints);
-            info.AddValue("Experience", Experience);
-            info.AddValue("Ability", Ability);
-            info.AddValue("MapId", MapId);
-            info.AddValue("MapX", MapX);
-            info.AddValue("AbilityExp", AbilityExp);
-            info.AddValue("BaseHp", BaseHp);
-            info.AddValue("BaseMp", BaseMp);
-            info.AddValue("Hp", Hp);
-            info.AddValue("Mp", Mp);
-            info.AddValue("BaseStr", BaseStr);
-            info.AddValue("BaseInt", BaseStr);
-            info.AddValue("BaseWis", BaseStr);
-            info.AddValue("BaseDex", BaseStr);
-            info.AddValue("BaseCon", BaseStr);
-            info.AddValue("Gold", Gold);
-            info.AddValue("IsMaster", IsMaster);
-            info.AddValue("Dead", Dead);
-            info.AddValue("Grouping", Grouping);
-            info.AddValue("PortraitData", PortraitData);
-            info.AddValue("ProfileText", ProfileText);
-            info.AddValue("LoginTime", LoginTime);
-            info.AddValue("LogoffTime", LogoffTime);
-            info.AddValue("UserFlags", UserFlags);
-            info.AddValue("PlayerStatus", Status);
-                
-            info.AddValue("LegendMarks", Legend);
-            info.AddValue("Inventory", Inventory, typeof(Inventory));
-            info.AddValue("Equipment", Equipment, typeof(Inventory));
-
-
-
-
-        }*/
     }
 }